# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AssetMind Backend API",
        "description": "REST API untuk Sistem Pendukung Keputusan Perawatan Aset Industri Perkebunan Kelapa Sawit. Backend Laravel terintegrasi dengan AI Engine (FastAPI) untuk embedding, chatbot RAG, dan rekomendasi pemeliharaan.",
        "contact": {
            "name": "AssetMind",
            "email": "firasyan.daffa123@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Server lokal (Laravel artisan serve)"
        }
    ],
    "paths": {
...
        "/api/assets": {
            "get": {
                "tags": [
                    "Assets"
                ],
                "summary": "List aset",
                "operationId": "assetsIndex",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerSanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Assets"
                ],
                "summary": "Buat aset baru (teknik)",
                "operationId": "assetsStore",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "kategori",
                                    "nomor_peralatan",
                                    "nama_mesin",
                                    "area_mesin",
                                    "maintenance_interval_hours"
                                ],
                                "properties": {
                                    "kategori": {
                                        "type": "string"
                                    },
                                    "nomor_peralatan": {
                                        "type": "string"
                                    },
                                    "nama_mesin": {
                                        "type": "string"
                                    },
                                    "area_mesin": {
                                        "type": "string"
                                    },
                                    "merek": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "tahun_pembelian": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "maintenance_interval_hours": {
                                        "type": "integer"
                                    },
                                    "foto_kondisi": {
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    },
                                    "manual_book": {
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerSanctum": []
                    }
                ]
            }
        },
        "/api/assets/{id}": {
            "get": {
                "tags": [
                    "Assets"
                ],
                "summary": "Detail aset",
                "operationId": "assetsShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail aset berhasil diambil"
                    },
                    "404": {
                        "description": "Aset tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerSanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Assets"
                ],
                "summary": "Perbarui aset (khusus teknik)",
                "description": "Memperbarui aset. Menggunakan multipart/form-data karena bisa mengunggah foto kondisi & manual book (PDF).",
                "operationId": "assetsUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "kategori",
                                    "nomor_peralatan",
                                    "nama_mesin",
                                    "area_mesin",
                                    "maintenance_interval_hours"
                                ],
                                "properties": {
                                    "kategori": {
                                        "type": "string",
                                        "example": "Mesin Produksi"
                                    },
                                    "nomor_peralatan": {
                                        "type": "string",
                                        "example": "P-10"
                                    },
                                    "nama_mesin": {
                                        "type": "string",
                                        "example": "Mesin Screw Press P-10"
                                    },
                                    "area_mesin": {
                                        "type": "string",
                                        "example": "Stasiun Press"
                                    },
                                    "merek": {
                                        "type": "string",
                                        "example": "Stork",
                                        "nullable": true
                                    },
                                    "tahun_pembelian": {
                                        "type": "integer",
                                        "example": 2020,
                                        "nullable": true
                                    },
                                    "maintenance_interval_hours": {
                                        "type": "integer",
                                        "example": 2000
                                    },
                                    "foto_kondisi": {
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    },
                                    "manual_book": {
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data aset berhasil diperbarui"
                    },
                    "422": {
                        "description": "Validasi gagal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerSanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Assets"
                ],
                "summary": "Hapus aset (khusus teknik)",
                "operationId": "assetsDestroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data aset berhasil dihapus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aset tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerSanctum": []
                    }
                ]
            }
        },
        "/api/assets/{asset}/qr-detail": {
            "get": {
                "tags": [
                    "Assets"
                ],
                "summary": "Detail aset dari QR code (publik)",
                "description": "Endpoint publik (tanpa autentikasi) yang ditautkan pada QR code aset. Mengembalikan profil aset + ringkasan jam operasi, histori pemeliharaan, dan histori perbaikan.",
                "operationId": "assetsQrDetail",
                "parameters": [
                    {
                        "name": "asset",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail aset dari QR code berhasil diambil"
                    },
                    "404": {
                        "description": "Aset tidak ditemukan"
                    }
                }
            }
        },
        "/api/assets/{asset}/qr-code": {
            "post": {
                "tags": [
                    "Assets"
                ],
                "summary": "Buat ulang QR code aset (khusus teknik)",
                "operationId": "assetsRegenerateQrCode",
                "parameters": [
                    {
                        "name": "asset",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "QR code aset berhasil dibuat ulang"
                    },
                    "404": {
                        "description": "Aset tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerSanctum": []
                    }
                ]
            }
        },
        "/api/assets/{id}/knowledge-summary": {
            "get": {
                "tags": [
                    "Assets"
                ],
                "summary": "Ringkasan knowledge aset",
                "description": "Mengembalikan jumlah knowledge document, knowledge chunks, laporan pemeliharaan, catatan perbaikan, dan jam operasi terbaru untuk satu aset.",
                "operationId": "assetsKnowledgeSummary",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ringkasan knowledge aset berhasil diambil"
                    },
                    "404": {
                        "description": "Aset tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerSanctum": []
                    }
                ]
            }
        },
...
        "/api/ingestion/assets/{asset}/profile": {
            "post": {
                "tags": [
                    "Knowledge"
                ],
                "summary": "Ingest profil aset ke knowledge base (khusus teknik)",
                "description": "Membuat embedding profil aset via AI Engine lalu menyimpannya ke pgvector.",
                "operationId": "ingestAssetProfile",
                "parameters": [
                    {
                        "name": "asset",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Profil aset berhasil di-ingest"
                    },
                    "500": {
                        "description": "Gagal ingest profil aset"
                    }
                },
                "security": [
                    {
                        "bearerSanctum": []
                    }
                ]
            }
        },
        "/api/ingestion/assets/{asset}/manual-book": {
            "post": {
                "tags": [
                    "Knowledge"
                ],
                "summary": "Ingest manual book aset (PDF) ke knowledge base (khusus teknik)",
                "description": "Mengekstrak teks PDF manual book, membuat embedding via AI Engine, lalu menyimpannya ke pgvector.",
                "operationId": "ingestAssetManualBook",
                "parameters": [
                    {
                        "name": "asset",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Manual book aset berhasil di-ingest"
                    },
                    "500": {
                        "description": "Gagal ingest manual book aset"
                    }
                },
                "security": [
                    {
                        "bearerSanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "MessageResponse": {
                "properties": {
                    "message": {
                        "description": "Definisi dasar dokumentasi OpenAPI/Swagger untuk AssetMind Backend API.\n\nMemakai PHP 8 Attributes (didukung swagger-php v6).\nAnotasi path tersebar di controller masing-masing.",
                        "type": "string",
                        "example": "Operasi berhasil"
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "ForbiddenError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Forbidden. Data ini bukan milik Anda."
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerSanctum": {
                "type": "http",
                "description": "Masukkan token Sanctum dari endpoint /auth/login (tanpa kata 'Bearer').",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}